+++ src/Reflection/FileResource.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Typhoon\Reflection;
 
 use Typhoon\Reflection\Exception\FileNotReadable;
-use Typhoon\Reflection\Metadata\ChangeDetector;
 
 /**
  * @api
@@ -23,8 +22,6 @@ final class FileResource
     public readonly string|false $extension;
 
     private ?string $contents = null;
-
-    private ?ChangeDetector $changeDetector = null;
 
     public function __construct(string $file, string|false $extension = false)
     {
@@ -51,17 +48,6 @@ final class FileResource
         }
 
         return $this->contents = $contents;
-    }
-
-    /**
-     * @deprecated will be removed in 0.4.0
-     * @throws FileNotReadable
-     */
-    public function changeDetector(): ChangeDetector
-    {
-        trigger_deprecation('typhoon/reflection', '0.3.1', 'Method %s() is deprecated and will be removed on 0.4.0.', __METHOD__);
-
-        return $this->changeDetector ??= ChangeDetector::fromFileContents($this->file, $this->contents());
     }
 
     public function isInternal(): bool