+++ src/Type/TypeVisitor.php
@@ -51,14 +51,6 @@ interface TypeVisitor
     public function classString(Type $self, Type $class): mixed;
 
     /**
-     * @deprecated will be removed in 0.4.0
-     * @param Type<non-empty-string> $self
-     * @param non-empty-string $class
-     * @return TReturn
-     */
-    public function classStringLiteral(Type $self, string $class): mixed;
-
-    /**
      * @param Type<\Closure> $self
      * @param list<Parameter> $parameters
      * @return TReturn
@@ -218,12 +210,6 @@ interface TypeVisitor
      * @return TReturn
      */
     public function union(Type $self, array $types): mixed;
-
-    /**
-     * @deprecated will be removed in 0.4.0
-     * @return TReturn
-     */
-    public function value(Type $self, Type $type): mixed;
 
     /**
      * @return TReturn