+++ src/Reflection/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0
+
+- Removed `FileResource::changeDetector()`.
+- Removed `NativeReflectionFileLocator`.
+- Removed `NativeReflectionLocator`.
+- Changed `ClassLocator::locateClass()` return type to `?FileResource`.
+
 ## 0.3.1
 
 - Implemented `getType()`, `getReturnType()`, `getTentativeReturnType()`, `hasTentativeReturnType()`, `hasReturnType()`.